fix(hazards): default displacement threshold to the populated floor so map/chart/dropdown agree
The threshold dropdown default read "3.0" while the map still filtered at the
deadband bound, so the map drew the ±1 "within uncertainty" band the chart hides.
Default the effective threshold to the per-type smallest data-populated edge (the
same value the dropdown offers as default) instead of the deadband bound, computed
at runtime from the SLD + live magnitudes — nothing hard-coded. Now the map, chart
and dropdown all agree, and the map hides the deadband by default like the chart.
The chart and KPIs are unchanged (they already excluded the deadband); only the map
tightens. Threshold math moved to a pure displacement-thresholds module shared by
the filter UI and the context. Verified against prod: Cumulative floor 3, Yearly
floor 2, both above their deadband. (#525)
