+++ src/routes/_map/hazards-review/-components/popups/displacement-layer-filters.tsx
@@ -24,7 +24,7 @@ function isDefaultDataQuals(excluded: ReadonlySet<string>): boolean {
         && DEFAULT_EXCLUDED_DATA_QUALS.every(q => excluded.has(q))
 }
 import { useDisplacementBasinsForType, useDisplacementBasinYearIndexForType, useDisplacementDataQualsForType, useDisplacementSldBins, useDisplacementValueMagnitudesForType, useDisplacementYearsForType } from './use-displacement-queries'
-import { getZeroBound, type SldBin } from './displacement-sld-legend'
+import { getPopulatedBinBoundaries } from './displacement-thresholds'
 
 // Label the year dropdown by type semantics: 'Water Year' for Yearly,
 // 'Period End Year' for Cumulative + Vertical Displacement Rate.
@@ -255,37 +255,6 @@ function DisplacementLayerFilters({ typeValue }: { typeValue: DisplacementType }
     )
 }
 
-// Positive SLD class edges (magnitudes), deduped + sorted ascending. These are
-// the structural boundaries from the style; `getPopulatedBinBoundaries` trims
-// them to the ones the data actually backs.
-function getBinBoundaries(bins: SldBin[]): number[] {
-    const edges = new Set<number>()
-    for (const b of bins) {
-        if (b.isZero) continue
-        for (const v of [b.min, b.max]) {
-            if (Number.isFinite(v)) edges.add(Math.abs(v))
-        }
-    }
-    return Array.from(edges).filter(v => v > 0).sort((a, b) => a - b)
-}
-
-// Threshold options: SLD class edges kept only when a real, non-deadband feature
-// sits in the band [edge, nextEdge). Two consecutive edges filter to the same
-// set unless some feature's |value| falls between them, so an SLD class the data
-// never fills would otherwise offer a redundant option. Concretely for
-// Cumulative, contours are odd inches and ±1 is the deadband, so the 1–3 in band
-// is empty and "1 in" would filter identically to "3 in" — this drops the 1.
-// `magnitudes` is the ascending distinct |value_inches| present for the type.
-export function getPopulatedBinBoundaries(bins: SldBin[], magnitudes: number[]): number[] {
-    const edges = getBinBoundaries(bins)
-    const zeroBound = getZeroBound(bins) ?? 0
-    const measured = magnitudes.filter(m => m > zeroBound)
-    return edges.filter((edge, i) => {
-        const next = edges[i + 1] ?? Infinity
-        return measured.some(m => m >= edge && m < next)
-    })
-}
-
 interface ThresholdSelectProps {
     typeValue: ChartedType
     currentThresholdIn: number
@@ -295,13 +264,12 @@ interface ThresholdSelectProps {
 
 // Threshold as a dropdown of the SLD bin edges the data actually fills (see
 // getPopulatedBinBoundaries — empty-band edges are dropped so no two options
-// filter to the same set). The first, smallest offered edge is marked
-// "· default"; picking it resets to the SLD default (the Zero-deadband bound),
-// which for the chart/stats collapses onto that edge. Larger edges tighten the
-// filter on |value| (both signs). Below-default values aren't offered — the
-// deadband is measurement noise, so reviewers can only raise the bar. (At the
-// default the map still draws the deadband band, since the map filters on the
-// effective bound, not this label — a pre-existing map/chart split.)
+// filter to the same set). The first, smallest offered edge is the per-type
+// default (marked "· default") and is exactly what the effective threshold
+// resolves to when unset, so the map, chart, and dropdown all agree. Picking it
+// resets to that default; larger edges tighten the filter on |value| (both
+// signs). Below the default isn't offered — that band is the measurement-noise
+// deadband, which the chart and (at the default) the map both exclude.
 function ThresholdSelect({ typeValue, currentThresholdIn, onChange, onReset }: ThresholdSelectProps) {
     const styleName = getStyleNameForType(typeValue) ?? ''
     const { data: sldBins = [] } = useDisplacementSldBins(styleName)
@@ -311,10 +279,10 @@ function ThresholdSelect({ typeValue, currentThresholdIn, onChange, onReset }: T
 
     const fmt = (n: number) => n.toFixed(1)
     const defaultThreshold = boundaries[0]
-    // The effective default is the SLD deadband bound, which can sit below the
-    // smallest offered edge (the deadband is excluded from the chart, so it
-    // collapses onto that edge). Clamp up so the control reflects the option
-    // actually in force instead of rendering blank against a value with no item.
+    // currentThresholdIn (the effective threshold) defaults to this same floor,
+    // so normally it equals defaultThreshold. Clamp up only so a stale sub-floor
+    // override (e.g. an old URL) can't render blank against options that start at
+    // the floor.
     const selected = fmt(Math.max(currentThresholdIn, defaultThreshold))
 
     return (
